Update 2013-10-statement-community.html.md
I changed the order so all the "positive" messages are there to be read first.
Background: If you start off with "no", "negative messages", telling the people what NOT to do etc. you are loosing exactly the audience that has behavior which is addressed in the statement. 

I added a bulletpoints for an overview. So sloppy TL;TR readers might get interested.